Classic algorithm's hierarchy changed
