+++ Test.py
@@ -7,30 +7,30 @@ canonical_factory = CanonicalFactory()
 genitor_factory = GenitorFactory()
 
 island_model = IslandModel([
-    canonical_factory.build_canonical(fitness_function=lambda x: sum(x) + 1, pop_size=100, gen_len=1000),
+    canonical_factory.build_canonical(fitness_function=lambda x: sum(x) + 1, pop_size=100, gen_len=500),
     genitor_factory.build_genitor(fitness_function=lambda x: sum(x)+1, recombinator="single_point_crossover",
-                                mutator="point_mutation", pop_size=100, gen_len=1000),
+                                mutator="point_mutation", pop_size=100, gen_len=500),
     genitor_factory.build_genitor(fitness_function=lambda x: sum(x)+1, recombinator="shuffler_crossover",
-                                mutator="point_mutation", pop_size=100, gen_len=1000)
+                                mutator="point_mutation", pop_size=100, gen_len=500)
 ], fitness_function=lambda x: sum(x) + 1)
 
-canonical = canonical_factory.build_canonical(fitness_function=lambda x: sum(x) + 1, pop_size=100, gen_len=1000)
+canonical = canonical_factory.build_canonical(fitness_function=lambda x: sum(x) + 1, pop_size=100, gen_len=500)
 
 
 genitor = genitor_factory.build_genitor(fitness_function=lambda x: sum(x)+1, recombinator="single_point_crossover",
-                                mutator="point_mutation", pop_size=100, gen_len=1000)
+                                mutator="point_mutation", pop_size=100, gen_len=500)
 
 def plot_transform(algorithm):
     return [[x[1] for x in algorithm._get_memory()], [x[0] for x in algorithm._get_memory()]]
 
 # Calculate
-print(f"Canonical: {canonical.evaluate(T=100000, p_mut=.9, p_gene_mut=.7)}")
-print(f"Genitor: {genitor.evaluate(T=500, p_mut=.7, p_gene_mut=.9)}")
+print(f"Canonical: {canonical.evaluate(T=100, p_mut=.9, p_gene_mut=.7)}")
+print(f"Genitor: {genitor.evaluate(T=100, p_mut=.7, p_gene_mut=.9)}")
 
-print("Island Model: " + str(island_model.evaluate(100,
-    [300, .5, .5],
-    [700, .9, .7],
-    [700, .9, .9])))
+print("Island Model: " + str(island_model.evaluate(10,
+    [100, .5, .5],
+    [100, .9, .7],
+    [100, .9, .9])))
 
 # Get convergence plot
 plt.subplot(121)
